+++ example/lib/main.dart
@@ -21,6 +21,9 @@ class AccordionApp extends StatelessWidget {
 
 class AccordionPage extends StatelessWidget //__
 {
+  final _headerStyle =
+      TextStyle(color: Colors.white, fontSize: 17, fontWeight: FontWeight.bold);
+
   build(context) => Scaffold(
         backgroundColor: Colors.blueGrey[100],
         appBar: AppBar(
@@ -28,62 +31,60 @@ class AccordionPage extends StatelessWidget //__
         ),
         body: Accordion(
           maxOpenSections: 1,
-          headerTextStyle: TextStyle(
-              color: Colors.white, fontSize: 17, fontWeight: FontWeight.bold),
           leftIcon: Icon(Icons.audiotrack, color: Colors.white),
           children: [
             AccordionSection(
               isOpen: true,
-              headerText: 'Introduction',
+              header: Text('Introduction', style: _headerStyle),
               content: Icon(Icons.airplanemode_active,
                   size: 200, color: Colors.amber),
             ),
             AccordionSection(
               isOpen: true,
-              headerText: 'About Us',
+              header: Text('About Us', style: _headerStyle),
               content: Icon(Icons.airline_seat_flat,
                   size: 120, color: Colors.blue[200]),
             ),
             AccordionSection(
               isOpen: true,
-              headerText: 'Company Info',
+              header: Text('Company Info', style: _headerStyle),
               content: Icon(Icons.airplay, size: 70, color: Colors.green[200]),
             ),
             AccordionSection(
               isOpen: true,
-              headerText: 'Contact',
+              header: Text('Contact', style: _headerStyle),
               content: Icon(Icons.contact_page, size: 300, color: Colors.grey),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Technical Jobs',
+              header: Text('Technical Jobs', style: _headerStyle),
               content: Icon(Icons.computer, size: 200, color: Colors.amber),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Administrative Jobs',
+              header: Text('Administrative Jobs', style: _headerStyle),
               content: Icon(Icons.emoji_people, size: 200, color: Colors.amber),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Culture',
+              header: Text('Culture', style: _headerStyle),
               content:
                   Icon(Icons.calculate_rounded, size: 200, color: Colors.green),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Community',
+              header: Text('Community', style: _headerStyle),
               content: Icon(Icons.commute_outlined,
                   size: 200, color: Colors.blueAccent),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Friends Of Us',
+              header: Text('Friends Of Us', style: _headerStyle),
               content: Icon(Icons.child_friendly, size: 200, color: Colors.red),
             ),
             AccordionSection(
               isOpen: false,
-              headerText: 'Map',
+              header: Text('Map', style: _headerStyle),
               content: Icon(Icons.map, size: 200, color: Colors.blue),
             ),
           ],